requirements.txt
openai>=1.66.0
python-dotenv>=1.0.1
opencv-python>=4.10.0.84
